implement snake and example with a tracking cam in wgpu window
