+++ examples/snake/Cargo.toml
@@ -7,6 +7,7 @@ hewn = { path = "../../" }
 wasm-bindgen = "0.2"
 js-sys = "0.3"
 rand = "0.8"
+winit = "0.30.12"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { version = "0.2", features = ["js"] }
