+++ src/main.rs
@@ -1,3 +1,4 @@
+use hewn::game::GameLogic;
 use hewn::runtime::{initialize_terminal_io, TerminalRuntime};
 use hewn::view::cursor::FollowPlayerYCursorStrategy;
 use hewn::view::{ScreenDimensions, TerminalRenderer, View, ViewCoordinate};
@@ -7,6 +8,53 @@ const SCREEN_WIDTH: u16 = 50;
 
 mod game {
 
+    pub struct GameController {
+        speed: f32,
+        is_up_pressed: bool,
+        is_down_pressed: bool,
+        is_left_pressed: bool,
+        is_right_pressed: bool,
+    }
+
+    impl GameController {
+        pub(crate) fn new(speed: f32) -> Self {
+            Self {
+                speed,
+                is_up_pressed: false,
+                is_down_pressed: false,
+                is_left_pressed: false,
+                is_right_pressed: false,
+            }
+        }
+
+        pub(crate) fn handle_key(&mut self, key: KeyCode, is_pressed: bool) -> bool {
+            // Log key events for debugging
+            println!(
+                "[DEBUG] handle_key: key={:?} is_pressed={}",
+                key, is_pressed
+            );
+            match key {
+                KeyCode::KeyW | KeyCode::ArrowUp => {
+                    self.is_up_pressed = is_pressed;
+                    true
+                }
+                KeyCode::KeyS | KeyCode::ArrowDown => {
+                    self.is_down_pressed = is_pressed;
+                    true
+                }
+                KeyCode::KeyA | KeyCode::ArrowLeft => {
+                    self.is_left_pressed = is_pressed;
+                    true
+                }
+                KeyCode::KeyD | KeyCode::ArrowRight => {
+                    self.is_right_pressed = is_pressed;
+                    true
+                }
+                _ => false,
+            }
+        }
+    }
+
     use hewn::{
         ecs::{
             self, CameraFollow, Components, EntityId, PositionComponent, RenderComponent,
@@ -15,11 +63,15 @@ mod game {
         game::GameLogic,
         runtime::Key,
     };
+    use winit::keyboard::KeyCode;
 
     pub struct MinimalGame {
         started: bool,
         pub ecs: ecs::ECS,
         pub player_entity_id: EntityId,
+
+        pub game_controller: GameController,
+        last_player_position: Option<(u16, u16)>,
     }
 
     impl MinimalGame {
@@ -46,24 +98,49 @@ mod game {
                 camera_follow: None,
             });
 
+            // Get initial player position for tracking
+            let last_player_position = ecs
+                .get_entity_by_id(player_entity_id)
+                .and_then(|e| e.components.position.as_ref().map(|p| (p.x, p.y)));
+
             MinimalGame {
                 started: false,
                 ecs,
                 player_entity_id,
+                game_controller: GameController::new(1.0),
+                last_player_position,
             }
         }
 
-        fn update_player_velocity(&mut self, key: Option<Key>) {
+        fn update_player_velocity(&mut self) {
+            let key = if self.game_controller.is_up_pressed {
+                Some(Key::Up)
+            } else if self.game_controller.is_down_pressed {
+                Some(Key::Down)
+            } else if self.game_controller.is_left_pressed {
+                Some(Key::Left)
+            } else if self.game_controller.is_right_pressed {
+                Some(Key::Right)
+            } else {
+                None
+            };
+            if let Some(key) = key {
+                println!("[DEBUG] update_player_velocity: key={:?}", key);
+            }
+
             let player_entity = self.ecs.get_entity_by_id_mut(self.player_entity_id);
             let Some(player_entity) = player_entity else {
+                println!("[DEBUG] update_player_velocity: player entity not found");
                 return;
             };
             let Some(velocity) = &mut player_entity.components.velocity else {
+                println!("[DEBUG] update_player_velocity: velocity component not found");
                 return;
             };
             let Some(key) = &key else {
                 velocity.x = 0;
                 velocity.y = 0;
+                // println!("[DEBUG] update_player_velocity: No key pressed, velocity set to (0,0)");
                 return;
             };
 
@@ -86,29 +163,73 @@ mod game {
                 }
                 _ => {}
             }
+            println!(
+                "[DEBUG] update_player_velocity: key={:?} velocity set to ({}, {})",
+                key, velocity.x, velocity.y
+            );
+        }
+
+        fn fix_player_velocity_to_zero(&mut self) {
+            let player_entity = self.ecs.get_entity_by_id_mut(self.player_entity_id);
+            let Some(player_entity) = player_entity else {
+                println!("[DEBUG] fix_player_velocity_to_zero: player entity not found");
+                return;
+            };
+            let Some(velocity) = &mut player_entity.components.velocity else {
+                println!("[DEBUG] fix_player_velocity_to_zero: velocity component not found");
+                return;
+            };
+            velocity.x = 0;
+            velocity.y = 0;
         }
     }
 
     impl GameLogic for MinimalGame {
         fn start_game(&mut self) {
             self.started = true;
+            println!("[DEBUG] Game started");
         }
 
-        fn next(&mut self, key: Option<Key>) {
+        fn next(&mut self) {
             if !self.started {
                 return;
             }
 
-            self.update_player_velocity(key);
+            // Track previous position for debug
+            let prev_position = self
+                .ecs
+                .get_entity_by_id(self.player_entity_id)
+                .and_then(|e| e.components.position.as_ref().map(|p| (p.x, p.y)));
+
+            self.update_player_velocity();
             let collisions = self.ecs.collision_pass();
             if collisions
                 .iter()
                 .flatten()
                 .any(|entity_id| entity_id == &self.player_entity_id)
             {
-                self.update_player_velocity(None);
+                println!(
+                    "[DEBUG] Collision detected for player entity {:?}. Stopping movement.",
+                    self.player_entity_id
+                );
+                self.fix_player_velocity_to_zero();
             }
             self.ecs.step();
+
+            // After step, check if position changed
+            let new_position = self
+                .ecs
+                .get_entity_by_id(self.player_entity_id)
+                .and_then(|e| e.components.position.as_ref().map(|p| (p.x, p.y)));
+
+            if let (Some(prev), Some(new)) = (prev_position, new_position) {
+                if prev != new {
+                    println!(
+                        "[DEBUG] Player entity {:?} moved from ({}, {}) to ({}, {})",
+                        self.player_entity_id, prev.0, prev.1, new.0, new.1
+                    );
+                }
+            }
         }
 
         fn ecs(&self) -> &ECS {
@@ -133,11 +254,33 @@ mod game {
                 position.x, position.y, start_game_str
             ))
         }
+
+        fn handle_key(&mut self, key: KeyCode, pressed: bool) -> bool {
+            let result = self.game_controller.handle_key(key, pressed);
+            if result {
+                println!(
+                    "[DEBUG] handle_key: Player entity {:?} key={:?} pressed={}",
+                    self.player_entity_id, key, pressed
+                );
+                // Also print current velocity and position for debugging
+                if let Some(player_entity) = self.ecs.get_entity_by_id(self.player_entity_id) {
+                    if let Some(velocity) = &player_entity.components.velocity {
+                        println!("[DEBUG] Player velocity: ({}, {})", velocity.x, velocity.y);
+                    }
+                    if let Some(position) = &player_entity.components.position {
+                        println!("[DEBUG] Player position: ({}, {})", position.x, position.y);
+                    }
+                }
+            }
+            result
+        }
     }
 }
 
 fn main() {
-    hewn::render::run().unwrap();
+    let mut game = game::MinimalGame::new();
+    game.start_game();
+    hewn::render::app::run(Box::new(game)).unwrap();
     let (stdout, stdin) = initialize_terminal_io();
 
     let mut view = View {
@@ -154,13 +297,14 @@ fn main() {
 
     let mut game = game::MinimalGame::new();
     let mut runtime = TerminalRuntime::new(stdin, &mut game, &mut view);
-    runtime.start();
+    // runtime.start();
 }
 
 #[cfg(test)]
 mod test {
     use crate::game;
     use hewn::{game::GameLogic, runtime::Key};
+    use winit::keyboard::KeyCode;
 
     #[test]
     fn test_player_move() {
@@ -169,7 +313,8 @@ mod test {
         assert!(player.is_some());
 
         game.start_game();
-        game.next(Some(Key::Down));
+        game.handle_key(KeyCode::ArrowDown, true);
+        game.next();
 
         let player = game.ecs.get_entity_by_id(game.player_entity_id);
         assert!(player.is_some());
@@ -190,7 +335,8 @@ mod test {
         assert!(player.is_some());
 
         game.start_game();
-        game.next(Some(Key::Up));
+        game.handle_key(KeyCode::ArrowUp, true);
+        game.next();
 
         let player = game.ecs.get_entity_by_id(game.player_entity_id);
         assert!(player.is_some());
