+++ src/render/happy-tree.png
