+++ src/render/shader.wgsl
@@ -15,13 +15,13 @@ var<uniform> camera: CameraUniform;
 
 struct VertexInput {
     @location(0) position: vec3<f32>,
-    @location(1) tex_coords: vec2<f32>,
-}
+    @location(1) color: vec3<f32>,
+};
 
 struct VertexOutput {
     @builtin(position) clip_position: vec4<f32>,
-    @location(0) tex_coords: vec2<f32>,
-}
+    @location(0) color: vec3<f32>,
+};
 
 @vertex
 fn vs_main(
@@ -35,18 +35,18 @@ fn vs_main(
         instance.model_matrix_3,
     );
     var out: VertexOutput;
-    out.tex_coords = model.tex_coords;
+    out.color = model.color;
     out.clip_position = camera.view_proj * model_matrix * vec4<f32>(model.position, 1.0);
     return out;
 }
 
-@group(0) @binding(0)
-var t_diffuse: texture_2d<f32>;
-@group(0) @binding(1)
-var s_diffuse: sampler;
+// @group(0) @binding(0)
+// var t_diffuse: texture_2d<f32>;
+// @group(0) @binding(1)
+// var s_diffuse: sampler;
 
 @fragment
 fn fs_main(in: VertexOutput) -> @location(0) vec4<f32> {
-    return textureSample(t_diffuse, s_diffuse, in.tex_coords);
+    return vec4<f32>(in.color, 1.0);
 }
